+++ adminer/lang/cs.inc.php
@@ -207,6 +207,7 @@ Lang::$translations = array(
 	'Add next' => 'Přidat další',
 	'Index Type' => 'Typ indexu',
 	'length' => 'délka',
+	'operator class' => 'třída operátorů',
 	'Algorithm' => 'Algoritmus',
 	'Condition' => 'Podmínka',