+++ adminer/lang/xx.inc.php
@@ -208,6 +208,7 @@ Lang::$translations = array(
 	'Add next' => 'Xx',
 	'Index Type' => 'Xx',
 	'length' => 'xx',
+	'operator class' => 'xx',
 	'Algorithm' => 'Xx',
 	'Condition' => 'Xx',